requirements.txt
pandas==2.2.3
matplotlib==3.10.1
fastnda==1.1.1
pyside6-essentials==6.10.0
pyarrow==23.0.0
openpyxl==3.1.5
pathvalidate==3.3.1